website/forms.py
from django import forms 
from website.models import Newsletter, Contact
from captcha.fields import CaptchaField



class NewsletterForm(forms.ModelForm):
    
    class Meta:
        model = Newsletter
        fields = '__all__'



class ContactForm(forms.ModelForm):
    captcha = CaptchaField()
    class Meta:
        model = Contact
        fields = '__all__'
