requirements.dev.txt
asgiref==3.8.1
Django==4.2.13
django-appconf==1.0.6
django-compressor==4.4
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.3.0
psycopg2==2.9.9
python-decouple==3.8
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.5.0
tzdata==2024.1
